Added: - dumping of arbitrary floats in GlobalVariablesDumper - dijet data members to GenDiphotons (should be eventually be a derived class)

+++ MicroAOD/interface/GlobalVariablesComputer.h
@@ -17,7 +17,7 @@ namespace flashgg {
         struct cache_t {
             float rho;
             int   nvtx;
-            unsigned int  event;
+            unsigned long long event;
             unsigned char lumi;
             unsigned int  run;
             float npu;
